refactor SingleOffice and created SingleOffice Details
